Make common ispyb callback for gridscans with no grid detect

+++ src/mx_bluesky/common/parameters/components.py
@@ -271,7 +271,7 @@ class XyzStarts(BaseModel):
 
 class OptionalGonioAngleStarts(BaseModel):
     # Gridscans have different omega starts
-    omega_starts_deg: list[float] = [0, 90]
+    omega_starts_deg: list[int] = [0, 90]
 
     phi_start_deg: float | None = None
     chi_start_deg: float | None = None
